+++ src/lib.rs
@@ -8,6 +8,7 @@
 pub mod ast;
 pub mod convert;
 pub mod diagnostic;
+pub mod editors;
 pub mod interpreter;
 pub mod learn;
 pub mod lexer;